fix: refactor naming convention
