+++ css/components/editor.css
@@ -1,5 +1,5 @@
 @import url("./keyframes.css");
-@import url("./codeeditor.css");
+@import url("./code-editor.css");
 @import url("./completion-messages.css");
 
 .editor-container {
