+++ js/description-formatter.js
